@@ -6,25 +6,28 @@
   "scripts": {
     "test": "mocha test/**/*.js"
   },
+  "type": "module",
   "author": "",
   "license": "ISC",
   "dependencies": {
     "bencode": "^2.0.1",
-    "lru-cache": "^5.1.1"
+    "lru-cache": "^10.1.0"
   },
   "optionalDependencies": {
-    "ed25519-supercop": "^1.2.0",
-    "sse4_crc32": "^6.0.1"
+    "ed25519-supercop": "^2.0.1",
+    "sse4_crc32": "^7.0.0"
   },
   "devDependencies": {
     "@types/bencode": "^2.0.0",
     "@types/debug": "^4.1.5",
-    "@types/lru-cache": "^5.1.0",
-    "@types/node": "^12.11.7",
-    "ed25519-supercop": "^1.2.0",
-    "mocha": "^6.2.2",
-    "sinon": "^7.5.0",
-    "sse4_crc32": "^6.0.1",
-    "typescript": "^3.6.4"
+    "@types/node": "^20.10.4",
+    "ed25519-supercop": "^2.0.1",
+    "mocha": "^10.2.0",
+    "sinon": "^17.0.1",
+    "sse4_crc32": "^7.0.0",
+    "typescript": "^5.3.3"
+  },
+  "imports": {
+    "#root/*": "./*.js"
   }
 }
